nlp-services/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
pymongo[srv]
python-dotenv